+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Sort Address Tokens [#909](https://github.com/EthVM/EthVM/pull/909)
 - Add in Address Contract Details [#914](https://github.com/EthVM/EthVM/pull/914)
 
+### Bug 
+- Hide state diff for failed tx [#913](https://github.com/EthVM/EthVM/pull/913)
+
 ===================================================================================
 
 ### Release v1.0.3-hotfix.2